Added confirmation prompt on step deletion.

Bug - https://devops.soroco.com/Soroco%20Products/Storyboard/_workitems/edit/8057

+++ src/app/components/createsop/right-panel/right-panel.component.ts
@@ -129,23 +129,32 @@ export class RightPanelComponent implements OnInit {
    * @param $event
    */
   onDeleteStep($event) {
-    if ($event.mode === "local") {
-      this.stepcontrolService.deleteStep($event.sectionIndex, $event.stepIndex);
-    } else if ($event.mode === "server") {
-      this.rightPanelService
-        .deleteStep(
-          this.pageService.userStoryId,
-          $event.stepId,
-          $event.insertionId,
-          $event.sectionInsertionId
-        )
-        .subscribe(res => {
+    this.confirm.confirmDelete(
+      "Are you sure you want to delete this step?",
+      () => {
+        if ($event.mode === "local") {
           this.stepcontrolService.deleteStep(
             $event.sectionIndex,
             $event.stepIndex
           );
-        });
-    }
+        } else if ($event.mode === "server") {
+          this.rightPanelService
+            .deleteStep(
+              this.pageService.userStoryId,
+              $event.stepId,
+              $event.insertionId,
+              $event.sectionInsertionId
+            )
+            .subscribe(res => {
+              this.stepcontrolService.deleteStep(
+                $event.sectionIndex,
+                $event.stepIndex
+              );
+            });
+        }
+      },
+      () => {}
+    );
   }
 
   /**
